fix(manipulation): broken tests
